+++ shakecast/app/orm/objects.py
@@ -1,19 +1,18 @@
 import os
 import sys
-import inspect as inspect_mod
 import time
 from math import floor
 from functools import wraps
-from werkzeug.security import generate_password_hash
 
-from sqlalchemy import *
-from sqlalchemy.ext.declarative import *
+from sqlalchemy import MetaData, Column, Integer, String, Float, ForeignKey, Table, select
+from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.ext.hybrid import hybrid_method, hybrid_property
-from sqlalchemy.orm import *
+from sqlalchemy.orm import relationship
 from sqlalchemy.sql import and_, or_
-from sqlalchemy.orm.session import Session as SessionClass
 
+from .engine import engine
 from ..util import *
+from .util import IMPACT_RANKS
 from ..impact import get_event_impact, get_impact
 
 # create a metadata object
@@ -24,6 +23,8 @@ Base = declarative_base(metadata=metadata)
 
 #######################################################################
 ########################## Facility Tables ############################
+
+
 class Facility(Base):
     __tablename__ = 'facility'
     shakecast_id = Column(Integer, primary_key=True)
@@ -59,21 +60,20 @@ class Facility(Base):
     metric = Column(String(20))
     updated = Column(Integer)
     updated_by = Column(String(32))
-    
-    shaking_history = relationship('FacilityShaking',
-                        backref='facility',
-                        cascade='save-update, delete, delete-orphan')
 
-    
+    shaking_history = relationship('FacilityShaking',
+                                   backref='facility',
+                                   cascade='save-update, delete, delete-orphan')
+
     aebm = relationship('Aebm',
                         backref='facility',
                         uselist=False,
                         cascade='save-update, delete, delete-orphan')
 
     attributes = relationship('Attribute',
-                        backref='facility',
-                        cascade='save-update, delete, delete-orphan')
-    
+                              backref='facility',
+                              cascade='save-update, delete, delete-orphan')
+
     def __repr__(self):
         return '''Facility(facility_id=%s,
                            facility_type=%s,
@@ -114,17 +114,16 @@ class Facility(Base):
                                             self.orange_beta,
                                             self.red_beta,
                                             self.metric)
-                           
+
     def __str__(self):
         return self.name
-    
+
     def get_attribute(self, name):
         for att in self.attributes:
             if att.name.lower() == name.lower():
                 return att.value
 
         return None
-
 
     def make_alert_level(self, shaking_point=None, shakemap=None):
         '''
@@ -135,28 +134,25 @@ class Facility(Base):
 
         return impact
 
-
     @hybrid_property
     def lat(self):
         return (self.lat_max + self.lat_min) / 2
-
 
     @hybrid_property
     def lon(self):
         return (self.lon_max + self.lon_min) / 2
 
- 
-    @hybrid_method    
+    @hybrid_method
     def in_grid(self, grid):
         # check if a point is within the boundaries of the grid
         return ((self.lon_min > grid.lon_min and
-                    self.lon_min < grid.lon_max and
-                    self.lat_min > grid.lat_min and
-                    self.lat_min < grid.lat_max) or
+                 self.lon_min < grid.lon_max and
+                 self.lat_min > grid.lat_min and
+                 self.lat_min < grid.lat_max) or
                 (self.lon_min > grid.lon_min and
                     self.lon_min < grid.lon_max and
                     self.lat_max > grid.lat_min and
-                    self.lat_max < grid.lat_max) or 
+                    self.lat_max < grid.lat_max) or
                 (self.lon_max > grid.lon_min and
                     self.lon_max < grid.lon_max and
                     self.lat_min > grid.lat_min and
@@ -172,7 +168,7 @@ class Facility(Base):
                 (self.lon_min < grid.lon_min and
                     self.lon_max > grid.lon_min and
                     self.lat_min < grid.lat_max and
-                    self.lat_max > grid.lat_max) or 
+                    self.lat_max > grid.lat_max) or
                 (self.lon_min < grid.lon_max and
                     self.lon_max > grid.lon_max and
                     self.lat_min < grid.lat_min and
@@ -181,39 +177,39 @@ class Facility(Base):
                     self.lon_max > grid.lon_max and
                     self.lat_min < grid.lat_min and
                     self.lat_max > grid.lat_max))
-    
+
     @in_grid.expression
     def in_grid(cls, grid):
         # check if a point is within the boundaries of the grid
-        return  or_(and_(cls.lon_min > grid.lon_min,
+        return or_(and_(cls.lon_min > grid.lon_min,
                         cls.lon_min < grid.lon_max,
                         cls.lat_min > grid.lat_min,
                         cls.lat_min < grid.lat_max),
-                    and_(cls.lon_min > grid.lon_min,
+                   and_(cls.lon_min > grid.lon_min,
                         cls.lon_min < grid.lon_max,
                         cls.lat_max > grid.lat_min,
                         cls.lat_max < grid.lat_max),
-                    and_(cls.lon_max > grid.lon_min,
+                   and_(cls.lon_max > grid.lon_min,
                         cls.lon_max < grid.lon_max,
                         cls.lat_min > grid.lat_min,
                         cls.lat_min < grid.lat_max),
-                    and_(cls.lon_max > grid.lon_min,
+                   and_(cls.lon_max > grid.lon_min,
                         cls.lon_max < grid.lon_max,
                         cls.lat_max > grid.lat_min,
                         cls.lat_max < grid.lat_max),
-                    and_(cls.lon_min < grid.lon_min,
+                   and_(cls.lon_min < grid.lon_min,
                         cls.lon_max > grid.lon_min,
                         cls.lat_min < grid.lat_min,
                         cls.lat_max > grid.lat_min),
-                    and_(cls.lon_min < grid.lon_min,
+                   and_(cls.lon_min < grid.lon_min,
                         cls.lon_max > grid.lon_min,
                         cls.lat_min < grid.lat_max,
                         cls.lat_max > grid.lat_max),
-                    and_(cls.lon_min < grid.lon_max,
+                   and_(cls.lon_min < grid.lon_max,
                         cls.lon_max > grid.lon_max,
                         cls.lat_min < grid.lat_min,
                         cls.lat_max > grid.lat_min),
-                    and_(cls.lon_min < grid.lon_max,
+                   and_(cls.lon_min < grid.lon_max,
                         cls.lon_max > grid.lon_max,
                         cls.lat_min < grid.lat_max,
                         cls.lat_max > grid.lat_max))
@@ -232,15 +228,17 @@ class Attribute(Base):
                         name={},
                         value={}
                     )'''.format(
-                                            self.shakecast_id,
-                                            self.facility_id,
-                                            self.name,
-                                            self.value
-                                   )
+            self.shakecast_id,
+            self.facility_id,
+            self.name,
+            self.value
+        )
+
 
 class Aebm(Base):
     __tablename__ = 'aebm'
-    facility_id = Column(Integer, ForeignKey('facility.shakecast_id'), primary_key=True)
+    facility_id = Column(Integer, ForeignKey(
+        'facility.shakecast_id'), primary_key=True)
     mbt = Column(String(20))
     sdl = Column(String(50))
     bid = Column(Integer)
@@ -267,7 +265,7 @@ class Aebm(Base):
         return (self.mbt and self.sdl and self.bid and
                 self.height and self.stories and self.year)
 
-    
+
 class FacilityShaking(Base):
     __tablename__ = 'facility_shaking'
     sort_by = 'weight'
@@ -304,7 +302,7 @@ class FacilityShaking(Base):
 
         # now let sqlalchemy do the real initialization
         super(FacilityShaking, self).__init__(**move_on)
-    
+
     @hybrid_property
     def exceedance_ratio(self):
         return getattr(self, self.alert_level, None)
@@ -318,25 +316,27 @@ class FacilityShaking(Base):
             if level['name'] == self.alert_level:
                 alert_level = level
                 break
-        
+
         for level in IMPACT_RANKS:
             if level['rank'] > alert_level['rank']:
                 if getattr(self.facility, level['name'], False):
                     next_alert_level = level
                     break
-        
+
         shaking = getattr(self, self.metric.lower())
         threshold = getattr(self.facility, alert_level['name'])
         if shaking and threshold:
             if next_alert_level:
-                next_threshold = getattr(self.facility, next_alert_level['name'])
+                next_threshold = getattr(
+                    self.facility, next_alert_level['name'])
 
-                exceedance = (shaking - threshold) / (next_threshold - threshold)
+                exceedance = (shaking - threshold) / \
+                    (next_threshold - threshold)
             elif alert_level:
                 exceedance = (shaking - threshold) / threshold
-    
+
         return exceedance
-    
+
     def __repr__(self):
         return '''FacilityShaking(shakemap_id=%s,
                                    facility_id=%s,
@@ -369,6 +369,7 @@ class FacilityShaking(Base):
                                                    self.psa03,
                                                    self.psa10,
                                                    self.psa30)
+
     def __cmp__(self, other):
         if int(getattr(self, self.sort_by.lower()) * 10000) > int(getattr(other, self.sort_by.lower()) * 10000):
             return 1
@@ -376,8 +377,8 @@ class FacilityShaking(Base):
             return -1
         else:
             return 0
-    
-    
+
+
 #######################################################################
 ######################## Notification Tables ##########################
 
@@ -391,7 +392,7 @@ class Notification(Base):
     status = Column(String(64))
     sent_timestamp = Column(Float)
     notification_file = Column(String(255))
-    
+
     def __repr__(self):
         return '''Notification(shakemap_id=%s,
                                group_id=%s,
@@ -399,13 +400,13 @@ class Notification(Base):
                                status=%s,
                                sent_timestamp=%s,
                                notification_file=%s)''' % (self.shakemap_id,
-                                                             self.group_id,
-                                                             self.notification_type,
-                                                             self.status,
-                                                             self.sent_timestamp,
-                                                             self.notification_file)
-    
-      
+                                                           self.group_id,
+                                                           self.notification_type,
+                                                           self.status,
+                                                           self.sent_timestamp,
+                                                           self.notification_file)
+
+
 class User(Base):
     __tablename__ = 'user'
     shakecast_id = Column(Integer, primary_key=True)
@@ -423,7 +424,7 @@ class User(Base):
     groups = relationship('Group',
                           secondary='user_group_connection',
                           backref='users')
-    
+
     def __repr__(self):
         return '''User(username=%s,
                        password=%s,
@@ -439,27 +440,25 @@ class User(Base):
                                            self.full_name,
                                            self.user_type)
 
-    @staticmethod      
+    @staticmethod
     def is_authenticated():
         return True
- 
+
     @staticmethod
     def is_active():
         return True
- 
+
     @staticmethod
     def is_anonymous():
         return False
- 
+
     def get_id(self):
         return unicode(self.shakecast_id)
-    
+
     def is_admin(self):
         return self.user_type.lower() == 'admin'
-    
-    
-  
-    
+
+
 class Group(Base):
     __tablename__ = 'group'
     shakecast_id = Column(Integer, primary_key=True)
@@ -472,19 +471,19 @@ class Group(Base):
     template = Column(String(255))
     updated = Column(Integer)
     updated_by = Column(String(32))
-    
+
     facilities = relationship('Facility',
                               secondary='facility_group_connection',
                               backref='groups')
-    
+
     past_notifications = relationship('Notification',
                                       backref='group',
                                       cascade='save-update, delete')
-    
+
     specs = relationship('GroupSpecification',
-                                  backref='group',
-                                  cascade='save-update, delete, delete-orphan')
-    
+                         backref='group',
+                         cascade='save-update, delete, delete-orphan')
+
     def __repr__(self):
         return '''Group(name=%s,
                         facility_type=%s,
@@ -497,21 +496,21 @@ class Group(Base):
                                           self.lon_max,
                                           self.lat_min,
                                           self.lat_max)
-                        
+
     def __str__(self):
         return self.name
-    
-    @hybrid_method    
+
+    @hybrid_method
     def in_grid(self, grid):
         # check if a point is within the boundaries of the grid
         return ((self.lon_min > grid.lon_min and
-                    self.lon_min < grid.lon_max and
-                    self.lat_min > grid.lat_min and
-                    self.lat_min < grid.lat_max) or
+                 self.lon_min < grid.lon_max and
+                 self.lat_min > grid.lat_min and
+                 self.lat_min < grid.lat_max) or
                 (self.lon_min > grid.lon_min and
                     self.lon_min < grid.lon_max and
                     self.lat_max > grid.lat_min and
-                    self.lat_max < grid.lat_max) or 
+                    self.lat_max < grid.lat_max) or
                 (self.lon_max > grid.lon_min and
                     self.lon_max < grid.lon_max and
                     self.lat_min > grid.lat_min and
@@ -527,7 +526,7 @@ class Group(Base):
                 (self.lon_min < grid.lon_min and
                     self.lon_max > grid.lon_min and
                     self.lat_min < grid.lat_max and
-                    self.lat_max > grid.lat_max) or 
+                    self.lat_max > grid.lat_max) or
                 (self.lon_min < grid.lon_max and
                     self.lon_max > grid.lon_max and
                     self.lat_min < grid.lat_min and
@@ -536,91 +535,92 @@ class Group(Base):
                     self.lon_max > grid.lon_max and
                     self.lat_min < grid.lat_min and
                     self.lat_max > grid.lat_max))
-        
-    @in_grid.expression   
+
+    @in_grid.expression
     def in_grid(cls, grid):
         # check if a point is within the boundaries of the grid
-        return  or_(and_(cls.lon_min > grid.lon_min,
+        return or_(and_(cls.lon_min > grid.lon_min,
                         cls.lon_min < grid.lon_max,
                         cls.lat_min > grid.lat_min,
                         cls.lat_min < grid.lat_max),
-                    and_(cls.lon_min > grid.lon_min,
+                   and_(cls.lon_min > grid.lon_min,
                         cls.lon_min < grid.lon_max,
                         cls.lat_max > grid.lat_min,
                         cls.lat_max < grid.lat_max),
-                    and_(cls.lon_max > grid.lon_min,
+                   and_(cls.lon_max > grid.lon_min,
                         cls.lon_max < grid.lon_max,
                         cls.lat_min > grid.lat_min,
                         cls.lat_min < grid.lat_max),
-                    and_(cls.lon_max > grid.lon_min,
+                   and_(cls.lon_max > grid.lon_min,
                         cls.lon_max < grid.lon_max,
                         cls.lat_max > grid.lat_min,
                         cls.lat_max < grid.lat_max),
-                    and_(cls.lon_min < grid.lon_min,
+                   and_(cls.lon_min < grid.lon_min,
                         cls.lon_max > grid.lon_min,
                         cls.lat_min < grid.lat_min,
                         cls.lat_max > grid.lat_min),
-                    and_(cls.lon_min < grid.lon_min,
+                   and_(cls.lon_min < grid.lon_min,
                         cls.lon_max > grid.lon_min,
                         cls.lat_min < grid.lat_max,
                         cls.lat_max > grid.lat_max),
-                    and_(cls.lon_min < grid.lon_max,
+                   and_(cls.lon_min < grid.lon_max,
                         cls.lon_max > grid.lon_max,
                         cls.lat_min < grid.lat_min,
                         cls.lat_max > grid.lat_min),
-                    and_(cls.lon_min < grid.lon_max,
+                   and_(cls.lon_min < grid.lon_max,
                         cls.lon_max > grid.lon_max,
                         cls.lat_min < grid.lat_max,
                         cls.lat_max > grid.lat_max))
-    
+
     @hybrid_method
     def point_inside(self, point):
         return (self.lat_min <= point.lat and
                 self.lat_max >= point.lat and
                 self.lon_min <= point.lon and
                 self.lon_max >= point.lon)
-    
+
     @point_inside.expression
     def point_inside(cls, point):
         return and_(cls.lat_min <= point.lat,
                     cls.lat_max >= point.lat,
                     cls.lon_min <= point.lon,
                     cls.lon_max >= point.lon)
-    
+
     def _get_specs(self,
-            notification_type,
-            scenario=False,
-            heartbeat=False,
-            inspection=None):
+                   notification_type,
+                   scenario=False,
+                   heartbeat=False,
+                   inspection=None):
         notification_type = notification_type.lower()
         specs = [s for s in self.specs if
-                s.notification_type.lower() == notification_type]
+                 s.notification_type.lower() == notification_type]
         filtered = [s for s in specs if
-                (((s.event_type.lower() == 'scenario') is scenario)
-                and ((s.event_type.lower() == 'heartbeat') is heartbeat)
-                or s.event_type.lower() == 'all')]
+                    (((s.event_type.lower() == 'scenario') is scenario)
+                     and ((s.event_type.lower() == 'heartbeat') is heartbeat)
+                     or s.event_type.lower() == 'all')]
 
         if len(filtered) > 0 and inspection is not None:
             filtered = [s for s in filtered if
-                    str(s.inspection_priority).lower() == inspection.lower()]
+                        str(s.inspection_priority).lower() == inspection.lower()]
 
         return filtered
 
     def get_new_event_spec(self, scenario=False, heartbeat=False):
         specs = self._get_specs(
-                'new_event', scenario=scenario, heartbeat=heartbeat)
+            'new_event', scenario=scenario, heartbeat=heartbeat)
 
         return specs[0] if len(specs) > 0 else None
 
     def get_inspection_spec(self, inspection, scenario=False):
-        specs = self._get_specs('damage', inspection=inspection, scenario=scenario)
+        specs = self._get_specs(
+            'damage', inspection=inspection, scenario=scenario)
 
         return specs[0] if len(specs) > 0 else None
 
     def gets_notification(self, notification_type, scenario=False, heartbeat=False):
         specs = self._get_specs(notification_type,
-                scenario=scenario,
-                heartbeat=heartbeat)
+                                scenario=scenario,
+                                heartbeat=heartbeat)
 
         return len(specs) > 0
 
@@ -669,9 +669,9 @@ class Group(Base):
             alert_level = notification.shakemap.get_alert_level()
             spec = self.get_inspection_spec(alert_level, scenario)
         else:
-            heartbeat = notification.event.type == 'heartbeat' 
+            heartbeat = notification.event.type == 'heartbeat'
             spec = self.get_new_event_spec(scenario=scenario,
-                    heartbeat=heartbeat)
+                                           heartbeat=heartbeat)
 
         if spec is None:
             return None
@@ -685,6 +685,7 @@ class Group(Base):
         # Return email as default
         return 'email'
 
+
 class GroupSpecification(Base):
     __tablename__ = 'group_specification'
     shakecast_id = Column(Integer, primary_key=True)
@@ -695,7 +696,7 @@ class GroupSpecification(Base):
     minimum_magnitude = Column(Integer)
     notification_format = Column(String(25))
     aggregate_name = Column(String(25))
-    
+
     def __repr__(self):
         return '''GroupSpecification(group_id=%s,
                                       notification_type=%s,
@@ -703,38 +704,38 @@ class GroupSpecification(Base):
                                       minimum_magnitude=%s,
                                       notification_format=%s,
                                       aggregate_name=%s)''' % (self.group_id,
-                                                         self.notification_type,
-                                                         self.inspection_priority,
-                                                         self.minimum_magnitude,
-                                                         self.notification_format,
-                                                         self.aggregate_name)
-    
+                                                               self.notification_type,
+                                                               self.inspection_priority,
+                                                               self.minimum_magnitude,
+                                                               self.notification_format,
+                                                               self.aggregate_name)
+
 
 # Connection tables
 user_group_connection = Table('user_group_connection', Base.metadata,
-    Column('user',
-           Integer,
-           ForeignKey('user.shakecast_id',
-                      ondelete='cascade'),
-           primary_key=True),
-    Column('group',
-           Integer,
-           ForeignKey('group.shakecast_id',
-                      ondelete='cascade'),
-           primary_key=True)
-)
+                              Column('user',
+                                     Integer,
+                                     ForeignKey('user.shakecast_id',
+                                                ondelete='cascade'),
+                                     primary_key=True),
+                              Column('group',
+                                     Integer,
+                                     ForeignKey('group.shakecast_id',
+                                                ondelete='cascade'),
+                                     primary_key=True)
+                              )
 
 facility_group_connection = Table('facility_group_connection', Base.metadata,
-    Column('facility',
-           Integer, ForeignKey('facility.shakecast_id',
-                               ondelete='cascade'),
-           primary_key=True),
-    Column('group',
-           Integer,
-           ForeignKey('group.shakecast_id',
-                      ondelete='cascade'),
-           primary_key=True)
-)
+                                  Column('facility',
+                                         Integer, ForeignKey('facility.shakecast_id',
+                                                             ondelete='cascade'),
+                                         primary_key=True),
+                                  Column('group',
+                                         Integer,
+                                         ForeignKey('group.shakecast_id',
+                                                    ondelete='cascade'),
+                                         primary_key=True)
+                                  )
 
 
 #######################################################################
@@ -755,13 +756,13 @@ class Event(Base):
     place = Column(String(255))
     time = Column(Integer)
     override_directory = Column(String(255))
-    
+
     shakemaps = relationship('ShakeMap',
                              backref='event')
-    
+
     notifications = relationship('Notification',
-                                    backref='event')
-    
+                                 backref='event')
+
     def __repr__(self):
         return '''Event(event_id=%s,
                         all_event_ids=%s,
@@ -778,10 +779,10 @@ class Event(Base):
                                        self.title,
                                        self.place,
                                        self.time)
-    
+
     def __str__(self):
         return self.event_id
-    
+
     @hybrid_property
     def directory_name(self):
         if self.override_directory:
@@ -792,7 +793,7 @@ class Event(Base):
             base_dir = get_data_dir()
 
         return os.path.join(base_dir, self.event_id)
-    
+
     @hybrid_property
     def local_products_dir(self):
         if self.override_directory:
@@ -809,28 +810,27 @@ class Event(Base):
         """
         Check if this Event is new
         """
-        
+
         # Get all ids associated with this event
         ids = self.all_event_ids.strip(',').split(',')
         events = []
-        
+
         for each_id in ids:
             stmt = (select([Event.__table__.c.event_id])
-                        .where(Event.__table__.c.event_id == each_id))
+                    .where(Event.__table__.c.event_id == each_id))
             result = engine.execute(stmt)
             events += [row for row in result]
-        
+
         return not bool(events)
-        
+
     def timestamp(self):
         """
         Non-static timestamp that changes based on the user's defined
         timezone
         """
-        from util import Clock
         clock = Clock()
         return (clock.from_time(self.time)
-                    .strftime('%Y-%m-%d %H:%M:%S'))
+                .strftime('%Y-%m-%d %H:%M:%S'))
 
 
 class ShakeMap(Base):
@@ -838,7 +838,7 @@ class ShakeMap(Base):
     shakecast_id = Column(Integer, primary_key=True)
     shakemap_id = Column(String(80))
     event_id = Column(Integer,
-            ForeignKey('event.shakecast_id'))
+                      ForeignKey('event.shakecast_id'))
     shakemap_version = Column(Integer)
     status = Column(String(64))
     type = Column(String(64))
@@ -861,22 +861,22 @@ class ShakeMap(Base):
     orage = Column(Integer)
     red = Column(Integer)
     override_directory = Column(String(255))
-    
+
     products = relationship('Product',
                             backref='shakemap',
                             cascade='save-update, delete')
-    
+
     notifications = relationship('Notification',
-                                 backref = 'shakemap')
+                                 backref='shakemap')
 
     facility_shaking = relationship('FacilityShaking',
                                     backref='shakemap',
                                     cascade='save-update, delete, delete-orphan')
-    
+
     local_products = relationship('LocalProduct',
-                            backref='shakemap',
-                            cascade='save-update, delete, delete-orphan')
-    
+                                  backref='shakemap',
+                                  cascade='save-update, delete, delete-orphan')
+
     def __repr__(self):
         return '''ShakeMap(shakemap_id=%s,
                            status=%s,
@@ -911,7 +911,7 @@ class ShakeMap(Base):
                                                     self.end_timestamp,
                                                     self.superceded_timestamp,
                                                     self.directory_name)
-    
+
     def __str__(self):
         return self.shakemap_id
 
@@ -930,7 +930,7 @@ class ShakeMap(Base):
             self.shakemap_id,
             self.shakemap_id + '-' + str(self.shakemap_version)
         )
-    
+
     @hybrid_property
     def local_products_dir(self):
         if self.override_directory:
@@ -940,44 +940,45 @@ class ShakeMap(Base):
             get_local_products_dir(),
             self.shakemap_id,
             self.shakemap_id + '-' + str(self.shakemap_version))
-    
+
     def old_maps(self):
         """
         Returns 0 for false and an integer count of old shakemaps for true
-        """        
+        """
         stmt = (select([ShakeMap.__table__.c.shakecast_id])
-                    .where(and_(ShakeMap.__table__.c.shakemap_id == self.shakemap_id,
-                                ShakeMap.__table__.c.shakemap_version < self.shakemap_version)))
-        
+                .where(and_(ShakeMap.__table__.c.shakemap_id == self.shakemap_id,
+                            ShakeMap.__table__.c.shakemap_version < self.shakemap_version)))
+
         result = engine.execute(stmt)
         old_shakemaps = [row for row in result]
-        
+
         return len(old_shakemaps)
 
     def is_new(self):
         stmt = (select([ShakeMap.__table__.c.shakecast_id])
-                    .where(and_(ShakeMap.__table__.c.shakemap_id == self.shakemap_id,
-                                ShakeMap.__table__.c.shakemap_version == self.shakemap_version)))
-        
+                .where(and_(ShakeMap.__table__.c.shakemap_id == self.shakemap_id,
+                            ShakeMap.__table__.c.shakemap_version == self.shakemap_version)))
+
         result = engine.execute(stmt)
         shakemaps = [row for row in result]
-        
+
         return len(shakemaps) == 0
-           
+
     def has_products(self, req_prods):
-        shakemap_prods = [prod.product_type for prod in self.products if prod.status == 'downloaded' and prod.error is None]
+        shakemap_prods = [
+            prod.product_type for prod in self.products if prod.status == 'downloaded' and prod.error is None]
         for prod in req_prods:
             if prod not in shakemap_prods:
                 return False
         return True
 
-    def get_map(self):       
+    def get_map(self):
         shakemap_file = os.path.join(self.directory_name, 'intensity.jpg')
         shakemap_image = open(shakemap_file, 'rb')
         map_image = shakemap_image.read()
         shakemap_image.close()
         return map_image
-    
+
     def get_overlay(self):
         overlay_names = ['intensity_overlay.png', 'ii_overlay.png']
         for name in overlay_names:
@@ -993,7 +994,7 @@ class ShakeMap(Base):
             alert_level = alert_levels[int(floor(insp_val))]
 
         return alert_level
-    
+
     def get_impact_summary(self):
         return get_event_impact(self.facility_shaking)
 
@@ -1003,15 +1004,16 @@ class ShakeMap(Base):
             self.begin_timestamp = time.time()
         else:
             self.superceded_timestamp = time.time()
-    
+
     def mark_processing_finished(self):
         self.status = 'processed'
         self.end_timestamp = time.time()
-    
+
     def sort_facility_shaking(self, sort_by, reverse=True):
         FacilityShaking.sort_by = sort_by
         self.facility_shaking = sorted(self.facility_shaking, reverse=reverse)
- 
+
+
 class Product(Base):
     __tablename__ = 'product'
     shakecast_id = Column(Integer, primary_key=True)
@@ -1029,7 +1031,7 @@ class Product(Base):
     error = Column(String(10000))
     update_username = Column(String(32))
     update_timestamp = Column(String(32))
-    
+
     def __repr__(self):
         return '''Product(product_type=%s,
                           name=%s,
@@ -1049,88 +1051,21 @@ class Product(Base):
                                                      self.update_username,
                                                      self.update_timestamp)
 
+
 class LocalProduct(Base):
     __tablename__ = 'local_products'
     id = Column(Integer, primary_key=True)
     type_id = Column(Integer,
-            ForeignKey('local_product_types.id'))
+                     ForeignKey('local_product_types.id'))
     shakemap_id = Column(Integer,
-            ForeignKey('shakemap.shakecast_id'))
+                         ForeignKey('shakemap.shakecast_id'))
     timestamp = Column(Integer)
 
     product_type = relationship('LocalProductType',
-                            backref='products')
+                                backref='products')
 
 
 class LocalProductType(Base):
     __tablename__ = 'local_product_types'
     id = Column(Integer, primary_key=True)
     type = Column(String(100))
-
-
-#######################################################################
-
-def clear_data(session, engine):
-    meta = MetaData(engine)
-    for table in reversed(meta.sorted_tables):
-        print 'Clear table %s' % table
-        session.execute(table.delete())
-    session.commit()
-
-# decorator for DB connection
-def dbconnect(func):
-    @wraps(func)
-    def inner(*args, **kwargs):
-        remove_session = False
-        session = None
-
-        # check if session is passed as arg
-        new_args = []
-        for arg in args:
-            if isinstance(arg, SessionClass):
-                session = arg
-            else:
-                new_args.append(arg)
-
-        # get session from kwargs -- will overwrite session from
-        # regular arg to ensure only one gets passed to the
-        # function
-        if 'session' in kwargs:
-            session = kwargs.pop('session')
-
-        # if there is no session, create one
-        if session is None:
-            session = Session()
-
-            # this session is created only for this function, destroy
-            # it on completion
-            remove_session = True
-
-        try:
-            # run the function
-            return_val = func(*new_args, session=session, **kwargs)
-            session.commit()
-        except:
-            session.rollback()
-            return_val = None
-            raise
-        finally:
-            refresh(return_val, session=session)
-
-            # function-specific session, close it
-            if remove_session is True:
-                session.expunge_all()
-                session.remove()
-
-        return return_val
-    return inner
-
-def refresh(obj, session=None):
-    if isinstance(obj, Base):
-        session.refresh(obj)
-    elif isinstance(obj, list):
-        for o in obj:
-            if isinstance(o, Base):
-                session.refresh(o)
-    elif isinstance(obj, dict):
-        pass
